third_party/WebKit/LayoutTests/media/track/track-cue-negative-timestamp.html

<!DOCTYPE html>
<title>Tests negative timestamps.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<video>
    <track id="testTrack" src="captions-webvtt/tc013-settings.vtt" default>
</video>
<script>
async_test(function(t) {
    var track = document.querySelector("track");

    track.onload = t.step_func_done(function() {
        // Test that cues with negative startTime are not added.
        assert_equals(testTrack.track.cues.length, 4);
        textCue = new VTTCue(-3439332606, 3.4, "Sausage?");
        testTrack.track.addCue(textCue);
        assert_equals(testTrack.track.cues.length, 4);

        // Test that cues with negative startTime and negative endTime are not added.
        assert_equals(testTrack.track.cues.length, 4);
        textCue = new VTTCue(-110, -3.4, "Pepperoni?");
        testTrack.track.addCue(textCue);
        assert_equals(testTrack.track.cues.length, 4);

        // Test that setting startTime and endTime to negative values does not affect the value.
        assert_equals(testTrack.track.cues[3].startTime, 121);
        testTrack.track.cues[3].startTime = -5;
        assert_equals(testTrack.track.cues[3].startTime, 121);
        assert_equals(testTrack.track.cues[3].endTime, 361200.5);
        testTrack.track.cues[3].endTime = -3439332606;
        assert_equals(testTrack.track.cues[3].endTime, 361200.5);
    });
});
</script>
